# specs/001-article-publish/contracts/api.yaml
openapi: 3.0.3
info:
  title: Hugo Admin Article Publishing API
  description: API endpoints for managing article publishing status
  version: 1.0.0
  contact:
    name: Hugo Admin Development

paths:
  /api/article/publish:
    post:
      summary: Publish a single article
      description: Changes article draft status from true to false
      operationId: publishArticle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_path
              properties:
                file_path:
                  type: string
                  description: Absolute path to article file
                  example: "/home/user/blog/content/post/my-article.md"
      responses:
        '200':
          description: Article published successfully or already published
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishResponse'
        '400':
          description: Bad request - invalid file path or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Article file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - article already being processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/article/publish/bulk:
    post:
      summary: Publish multiple articles
      description: Changes draft status to false for multiple articles
      operationId: bulkPublishArticles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - file_paths
              properties:
                file_paths:
                  type: array
                  items:
                    type: string
                  description: List of absolute file paths to publish
                  example: [
                    "/home/user/blog/content/post/article1.md",
                    "/home/user/blog/content/post/article2.md"
                  ]
      responses:
        '200':
          description: Bulk publish operation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkPublishResponse'
        '400':
          description: Bad request - invalid file paths or parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/article/status:
    get:
      summary: Get publish status of an article
      description: Returns current draft status and publishability information
      operationId: getArticleStatus
      parameters:
        - name: file_path
          in: query
          required: true
          description: Absolute path to article file
          schema:
            type: string
          example: "/home/user/blog/content/post/my-article.md"
      responses:
        '200':
          description: Article status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleStatusResponse'
        '400':
          description: Bad request - invalid file path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Article file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/article/status/bulk:
    get:
      summary: Get publish status of multiple articles
      description: Returns current draft status for multiple articles
      operationId: getBulkArticleStatus
      parameters:
        - name: file_paths
          in: query
          required: true
          description: Comma-separated list of absolute file paths
          schema:
            type: string
          example: "/home/user/blog/content/post/article1.md,/home/user/blog/content/post/article2.md"
      responses:
        '200':
          description: Article statuses retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkStatusResponse'
        '400':
          description: Bad request - invalid file paths
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PublishResponse:
      type: object
      required:
        - success
        - message
        - operation_id
      properties:
        success:
          type: boolean
          description: Whether the publish operation was successful
        message:
          type: string
          description: Human-readable result message
        operation_id:
          type: string
          format: uuid
          description: Unique identifier for this operation
        article_path:
          type: string
          description: Path to the published article
        draft_status_changed:
          type: boolean
          description: Whether the draft status was actually changed
        published_at:
          type: string
          format: date-time
          description: Timestamp when article was published
      example:
        success: true
        message: "Article published successfully"
        operation_id: "550e8400-e29b-41d4-a716-446655440000"
        article_path: "/home/user/blog/content/post/my-article.md"
        draft_status_changed: true
        published_at: "2025-11-14T10:30:00Z"

    BulkPublishResponse:
      type: object
      required:
        - success
        - total_count
        - published_count
        - failed_count
        - results
      properties:
        success:
          type: boolean
          description: Whether the bulk operation completed successfully
        total_count:
          type: integer
          description: Total number of articles processed
        published_count:
          type: integer
          description: Number of articles successfully published
        failed_count:
          type: integer
          description: Number of articles that failed to publish
        operation_id:
          type: string
          format: uuid
          description: Unique identifier for this bulk operation
        results:
          type: array
          items:
            $ref: '#/components/schemas/BulkPublishResult'
        errors:
          type: array
          items:
            type: string
          description: Global errors that affected multiple articles
      example:
        success: true
        total_count: 3
        published_count: 2
        failed_count: 1
        operation_id: "550e8400-e29b-41d4-a716-446655440001"
        results:
          - file_path: "/home/user/blog/content/post/article1.md"
            success: true
            published_at: "2025-11-14T10:30:00Z"
          - file_path: "/home/user/blog/content/post/article2.md"
            success: true
            published_at: "2025-11-14T10:30:01Z"
          - file_path: "/home/user/blog/content/post/article3.md"
            success: false
            error: "Permission denied"

    BulkPublishResult:
      type: object
      required:
        - file_path
        - success
      properties:
        file_path:
          type: string
          description: Path to the article
        success:
          type: boolean
          description: Whether this article was published successfully
        published_at:
          type: string
          format: date-time
          description: Timestamp when article was published (if successful)
        error:
          type: string
          description: Error message if publish failed

    ArticleStatusResponse:
      type: object
      required:
        - success
        - status
      properties:
        success:
          type: boolean
          description: Whether status retrieval was successful
        status:
          $ref: '#/components/schemas/ArticleStatus'
      example:
        success: true
        status:
          file_path: "/home/user/blog/content/post/my-article.md"
          is_draft: true
          is_publishable: true
          last_published: null
          publish_errors: []

    BulkStatusResponse:
      type: object
      required:
        - success
        - statuses
      properties:
        success:
          type: boolean
          description: Whether status retrieval was successful
        statuses:
          type: array
          items:
            $ref: '#/components/schemas/ArticleStatus'
      example:
        success: true
        statuses:
          - file_path: "/home/user/blog/content/post/article1.md"
            is_draft: false
            is_publishable: true
            last_published: "2025-11-14T10:30:00Z"
            publish_errors: []
          - file_path: "/home/user/blog/content/post/article2.md"
            is_draft: true
            is_publishable: true
            last_published: null
            publish_errors: []

    ArticleStatus:
      type: object
      required:
        - file_path
        - is_draft
        - is_publishable
      properties:
        file_path:
          type: string
          description: Path to the article
        is_draft:
          type: boolean
          description: Whether article is currently in draft state
        is_publishable:
          type: boolean
          description: Whether article can be published (no blocking issues)
        last_published:
          type: string
          format: date-time
          description: Last time article was published (null if never published)
        publish_errors:
          type: array
          items:
            type: string
          description: Issues preventing publication (empty if publishable)

    ErrorResponse:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          description: Human-readable error message
        error_code:
          type: string
          description: Machine-readable error code
        details:
          type: object
          description: Additional error details
      example:
        success: false
        error: "File not found"
        error_code: "FILE_NOT_FOUND"
        details:
          file_path: "/home/user/blog/content/post/nonexistent.md"

  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []